requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
django-filter==23.2
coverage
pytest
pytest-django
python-dotenv
